requirements.txt
Django==4.2.11
djangorestframework==3.15.0
djangorestframework-simplejwt==5.2.2
django-environ==0.12.0
psycopg2-binary==2.9.10
openpyxl==3.1.2
python-dotenv==1.0.0
